gradio-hello-world/requirements.txt
gradio
pypdf
google.generativeai